[TableFragmentation] Fix collapsed borders near fragmentation boundaries
This does some fixes to collapsed borders around fragmentation
boundaries.

Firstly it excludes the over/under border from the joint calculation
when at a fragmentation boundary. This prevents the border seemingly
having a random joint color when fragmentation is present.

We check if we should draw a "half" border if we are at a fragmentation
boundary, and we know we have previous/next table content. We check
this by looking at the row indices to determine if there are subsequent
rows.

I've made the test .tentative as the joint logic isn't really written
down anywhere, and everyone has bad behaviour here.

Bug: 1078927
Change-Id: I7a12458bb128f3a6650b2a408cef218ae1093dd3