History rewrite: Add WPT for srcdoc policy container inheritance
This CL adds a test for storing the policy container in the history
entry for about:srcdoc documents. The HTML standard requires that the
policy container is inherited fresh from the parent for e.g., history
navigations to about:srcdoc documents, as per [1].

Chrome and Firefox both fail this test, and apparently store the policy
container on the history entry, meaning for about:srcdoc documents, the
CSPs can get out-of-date from the parent.

[1]: https://html.spec.whatwg.org/multipage/origin.html#requires-storing-the-policy-container-in-history

R=domenic@chromium.org

Bug: 1329190
Change-Id: I87241b0f65c8ab6f1c58df866490c09dca80c915