Add popup support for <input>s styled as buttons and text fields
With this CL, several <input> element types, in addition to the existing
support for <button>, will be able to use the invoking attributes
(togglepopup, showpopup, and hidepopup) to invoke a popup. This CL
also adds enforcement of the resolution below, so that any button
that would otherwise submit a form cannot also/instead trigger a
popup.

https://github.com/openui/open-ui/issues/409#issuecomment-973176547

Bug: 1307772
Change-Id: I7eb5cd726bcacd26de5085871d7c3077c1f78baf